Add files via upload
The dataset is basically about the sales of different products sold in different cities of the US in the year 2019 and 2020. Here in the first visual, I used a slicer through which the data for selected cities would be shown. The second one is card visual that shows the sum of months sold the total products. A visual showing the ratio of the sum of quantity and sum of sales is also being presented which is in the pie chart visual. A line chart visual represents the graph for sum of sales made by year, month and days of year 2019 and 2020. My fifth visual is of donut chart that shows the sum of price each product has where it shows both the product's data value and percentage of prices. My last visual is of line and stacked column chart where I represent the sum of quantity ordered by the products and here it visuals the selling of each product.
The conclusion I found in this dataset is that AAA batteries has a huge amount of being sold in every city of US where LG dryer is the least selling product. The ratio of every product by the year 2019 and 2020 varies on the selling and in January 2020 it found a great loss in products selling.